+++ sctoolbox/utils/checker.py
@@ -4,6 +4,7 @@ import matplotlib
 import numpy as np
 import gzip
 import shutil
+import warnings
 
 import sctoolbox.utils as utils
 
@@ -299,3 +300,39 @@ def is_regex(regex):
 
     except re.error:
         return False
+
+
+def check_marker_lists(adata, marker_dict):
+    """
+    Remove genes in custom marker genes lists which are not present in dataset.
+
+    Parameters
+    ----------
+    adata : AnnData object
+        The anndata object containing features to annotate.
+    marker_dict : dict
+        Dictionary containing list of markergenes as values and coresponding cell types as key.
+
+    Returns
+    -------
+    Dictionary containing list of markergenes as values and coresponding cell types as key.
+    """
+    remove_key = []
+    for key, genes in marker_dict.items():
+        found_in_var = list()
+        not_found_in_var = list()
+        for g in genes:
+            if g in adata.var.index:
+                found_in_var.append(g)
+            else:
+                not_found_in_var.append(g)
+        if not found_in_var:
+            warnings.warn(f"No marker in {key} marker list can be found in the data. " +
+                          "Please check your marker list. Removing empty marker list form dictionary.")
+            remove_key.append(key)
+        elif not_found_in_var:
+            marker_dict[key] = found_in_var
+            print(f"Removed {not_found_in_var} from {key} marker gene list")
+    for i in remove_key:
+        marker_dict.pop(i)
+    return marker_dict