+++ CHANGES.rst
@@ -1,6 +1,7 @@
 0.2 (in progress)
 --------------------
-- Added parameters for the TOBIAS flags in the config file to write_TOBIAS_config()
+- Added parameters for the TOBIAS flags in the config file to write_TOBIAS_config()^
+- Added check for marker gene lists
 
 0.1.1 (24-05-2023)
 --------------------